fight-tracker connects to radar-collector w/ basic authentication
The flight-tracker app uses basic authentication to connect to
radar-collector

Using flight-tracker will now require a username and password to view
the index.html The username is user and the password is secure random
generated and output to the console at startup time.

Even though credentials are required for http, security can be
bypassed by connecting directly to websocket. See
FlightTrackerApplicationTests.